difftest runner: cleanup wgpu runner, remove variants of runners, use shader passthrough for rust-gpu
