+++ tests/difftests/tests/lang/core/intrinsics/black_box_noop/with-black-box/src/main.rs
@@ -1,9 +1,13 @@
 use difftest::config::{Config, TestMetadata};
-use difftest::scaffold::compute::{RustComputeShader, WgpuComputeTest};
+use difftest::scaffold::compute::{RustComputeShader, WgpuComputeTestMultiBuffer};
 
 fn main() {
     let config = Config::from_path(std::env::args().nth(1).unwrap()).unwrap();
-    let test = WgpuComputeTest::new(RustComputeShader::default(), [1, 1, 1], 12 * 4);
+    let test = WgpuComputeTestMultiBuffer::new_single_buffer(
+        RustComputeShader::default(),
+        [1, 1, 1],
+        12 * 4,
+    );
     test.run_test(&config).unwrap();
     config
         .write_metadata(&TestMetadata::u32())
