+++ tests/difftests/tests/lib/src/scaffold/compute/mod.rs
@@ -1,10 +1,10 @@
-mod ash;
+mod ash_runner;
 mod backend;
-mod wgpu;
+mod wgpu_runner;
 
 pub use crate::scaffold::shader::*;
-pub use ash::AshBackend;
+pub use ash;
+pub use ash_runner::AshBackend;
 pub use backend::{BufferConfig, BufferUsage, ComputeBackend, ComputeShaderTest, ComputeTest};
-pub use wgpu::{
-    WgpuBackend, WgpuComputeTest, WgpuComputeTestMultiBuffer, WgpuComputeTestPushConstants,
-};
+pub use wgpu;
+pub use wgpu_runner::WgpuComputeTestMultiBuffer;
